streamlit-app/requirements.txt
numpy
pandas
plotly.express
streamlit
kafka-python
prometheus_client